agent/lib/supabase_client.py
from supabase import create_client, Client
from dotenv import load_dotenv
import os

load_dotenv()


def get_supabase_client() -> Client:
    url = os.getenv("SUPABASE_URL")
    key = os.getenv("SUPABASE_SERVICE_ROLE_KEY")

    if not url or not key:
        raise EnvironmentError(
            "SUPABASE_URL and SUPABASE_SERVICE_ROLE_KEY "
            "must be set in environment."
        )

    return create_client(url, key)
