agent/requirements.txt
anthropic==0.49.0
supabase==2.15.0
fastapi==0.115.0
uvicorn==0.34.0
python-dotenv==1.0.1
pydantic==2.11.0
langgraph==0.3.0
langchain-anthropic==0.3.0
langchain-core==0.3.0
pytest==8.3.0
pytest-asyncio==0.25.0
httpx==0.28.0
